WIP: query morton code intervals
